requirements.txt
jsonpickle==3.3.0
telebot==0.0.5
python-dotenv==1.0.1
aiohttp==3.10.5